finished refactoring existing tests to test against generators directly.
