+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/apirequest/binders/SortGenerator.java
@@ -34,6 +34,52 @@ public interface SortGenerator {
     String DATE_TIME_STRING = "dateTime";
 
     /**
+     * Default implementation of {@link SortGenerator}.
+     */
+    SortGenerator DEFAULT_SORT_GENERATOR = new DefaultSortGenerator();
+
+    /**
+     * Utility method to parse the sort parameter from the api request into a mapping of data identifier to sort
+     * direction.
+     *
+     * @param sorts  The raw sorts parameter from the api request.
+     * @return the mapping of data identifier to sort direction.
+     */
+    static LinkedHashMap<String, SortDirection> generateSortDirectionMap(String sorts) {
+        LinkedHashMap<String, SortDirection> sortDirectionMap = new LinkedHashMap<>();
+
+        if (sorts != null && !sorts.isEmpty()) {
+            Arrays.stream(sorts.split(","))
+                    .map(e -> Arrays.asList(e.split("\\|")))
+                    .forEach(e -> sortDirectionMap.put(e.get(0), getSortDirection(e)));
+            return sortDirectionMap;
+        } else {
+            return null;
+        }
+    }
+
+    /**
+     * Utility method for resolving a pair of data identifier and sort direction from the api request sort parameter
+     * into a {@link SortDirection}. If not sort direction is specified, {@link SortDirection#DESC} is returned.
+     *
+     * @param columnWithDirection  The combination of a data identifier and the string version of the metric.
+     * @return the sort direction.
+     * @incubating it is likely the contract on this method will be cleaned up to make it more suitable for a public
+     * API.
+     */
+    @Incubating
+    static SortDirection getSortDirection(List<String> columnWithDirection) {
+        try {
+            return columnWithDirection.size() == 2 ?
+                    SortDirection.valueOf(columnWithDirection.get(1).toUpperCase(Locale.ENGLISH)) :
+                    SortDirection.DESC;
+        } catch (IllegalArgumentException ignored) {
+            String sortDirectionName = columnWithDirection.get(1);
+            throw new BadApiRequestException(SORT_DIRECTION_INVALID.format(sortDirectionName));
+        }
+    }
+
+    /**
      * Generates a sort that includes dateTime from an already generated mapping of sort data identifier to sort
      * direction. A utility for generating this mapping is provided on this interface as
      * {@link SortGenerator#generateSortDirectionMap(String)}. Implementors looking to expand the sort capability of
@@ -83,51 +129,4 @@ public interface SortGenerator {
             LinkedHashSet<LogicalMetric> logicalMetrics,
             MetricDictionary metricDictionary
     );
-
-    /**
-     * Utility method to parse the sort parameter from the api request into a mapping of data identifier to sort
-     * direction.
-     *
-     * @param sorts  The raw sorts parameter from the api request.
-     * @return the mapping of data identifier to sort direction.
-     */
-    static LinkedHashMap<String, SortDirection> generateSortDirectionMap(String sorts) {
-        LinkedHashMap<String, SortDirection> sortDirectionMap = new LinkedHashMap<>();
-
-        if (sorts != null && !sorts.isEmpty()) {
-            Arrays.stream(sorts.split(","))
-                    .map(e -> Arrays.asList(e.split("\\|")))
-                    .forEach(e -> sortDirectionMap.put(e.get(0), getSortDirection(e)));
-            return sortDirectionMap;
-        } else {
-            return null;
-        }
-    }
-
-    /**
-     * Utility method for resolving a pair of data identifier and sort direction from the api request sort parameter
-     * into a {@link SortDirection}. If not sort direction is specified, {@link SortDirection#DESC} is returned.
-     *
-     * @param columnWithDirection  The combination of a data identifier and the string version of the metric.
-     * @return the sort direction.
-     * @incubating it is likely the contract on this method will be cleaned up to make it more suitable for a public
-     * API.
-     */
-    @Incubating
-    static SortDirection getSortDirection(List<String> columnWithDirection) {
-        try {
-            return columnWithDirection.size() == 2 ?
-                    SortDirection.valueOf(columnWithDirection.get(1).toUpperCase(Locale.ENGLISH)) :
-                    SortDirection.DESC;
-        } catch (IllegalArgumentException ignored) {
-            String sortDirectionName = columnWithDirection.get(1);
-            throw new BadApiRequestException(SORT_DIRECTION_INVALID.format(sortDirectionName));
-        }
-    }
-
-    /**
-     * Default implementation of {@link SortGenerator}.
-     */
-    SortGenerator DEFAULT_SORT_GENERATOR = new DefaultSortGenerator();
-
 }
