+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/apirequest/binders/DefaultGranularityGeneratorSpec.groovy
@@ -1,6 +1,6 @@
 // Copyright 2018 Oath Inc.
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
-package com.yahoo.bard.webservice.web.apirequest.building
+package com.yahoo.bard.webservice.web.apirequest.binders
 
 import static com.yahoo.bard.webservice.data.time.DefaultTimeGrain.DAY
